move compare datasets init to util; code clean up
